% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{darken}
\alias{darken}
\title{darken}
\usage{
darken(color, factor = 1.4)
}
\arguments{
\item{color}{a character vector of valid R colors as names or hex}

\item{factor}{the factor by which the color should be darkened}
}
\value{
a vector of hex strings representing darkened colors
}
\description{
darken
}
\examples{
darken(c("white", "#565656"), 1.4)

}
