% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tetxpath.R
\name{geom_textpath}
\alias{geom_textpath}
\title{Create text along a path}
\usage{
geom_textpath(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  start_proportion = 0.45,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{start_proportion}{how far along the path should the text start?
This is useful for situations with longer paths and shorter strings.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are often
aesthetics, used to set an aesthetic to a fixed value, like \code{color =
"red"} or \code{size = 3}. They may also be parameters to the paired
 geom/stat.}
}
\description{
Allows text to follow a curved path in ggplot
}
\examples{
spiral <- data.frame(x = rev(sin(seq(0, 5*pi, length.out = 1000)) * 1000:1),
                     y = rev(cos(seq(0, 5*pi, length.out = 1000)) * 1000:1),
                     s = seq(1, 10, length.out = 1000),
                     z = paste("Like a circle in a spiral, like a",
                               "wheel within a wheel, never ending",
                               "or beginning on an ever spinning reel"))

ggplot(spiral, aes(x, y, label = z)) +
  geom_textpath(start_proportion = 0.2, size = 5.9, vjust=-0.4) +
  geom_path(color = "gray90") +
  coord_equal() +
  theme_void()


df <- data.frame(x = 1:1000, y = 1, z = "This is a perfectly flat label")

p <- ggplot(df, aes(x, y, label = z)) +
     geom_textpath(start_proportion = 0.25, size = 6)

p

p + coord_polar()
}
