% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomtextpath-package.R
\docType{package}
\name{geomtextpath-package}
\alias{geomtextpath}
\alias{geomtextpath-package}
\title{geomtextpath: Use Curved Text In Ggplot2}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Allows text to follow curved paths in ggplot2. This allows for neat labelling of curves and of plots in polar co-ordinates.
}
\author{
\strong{Maintainer}: Allan Cameron \email{Allan.Cameron@nhs.scot}

Authors:
\itemize{
  \item Teun van den Brand \email{tahvdbrand@gmail.com} (\href{https://orcid.org/0000-0002-9335-7468}{ORCID})
}

}
\keyword{internal}
