% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_textpath.R, R/geom_labelpath.R,
%   R/geom_textsegment.R, R/geom_textcurve.R
\docType{data}
\name{GeomTextpath}
\alias{GeomTextpath}
\alias{GeomLabelpath}
\alias{GeomTextsegment}
\alias{GeomLabelsegment}
\alias{GeomTextcurve}
\alias{GeomLabelcurve}
\title{The Geom object for a textpath}
\description{
This is the \code{ggproto} class that creates the textpath layer. It is not
intended to be used directly by the end user.
}
\keyword{datasets}
