% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomtextpath-package.R
\name{sibling_layers}
\alias{sibling_layers}
\title{Sibling layers}
\description{
The goal of \pkg{geomtextpath} is to label (curved) lines in a plot. The
\pkg{ggplot2} package has various ways to construct lines. For several of
the \pkg{ggplot2} line functions, there is a plain 'text' sibling and a
'label' sibling that includes a text box. Below is an overview of how
function in \pkg{geomtextpath} relate to those in \pkg{ggplot2}.
}
\details{
\tabular{lll}{
   \pkg{ggplot2} geom \tab Text equivalent \tab Label equivalent \cr
   \code{\link[ggplot2:geom_path]{geom_path()}} \tab \code{\link[=geom_textpath]{geom_textpath()}} \tab \code{\link[=geom_labelpath]{geom_labelpath()}} \cr
   \code{\link[ggplot2:geom_path]{geom_line()}} \tab \code{\link[=geom_textline]{geom_textline()}} \tab \code{\link[=geom_labelline]{geom_labelline()}} \cr
   \code{\link[ggplot2:geom_segment]{geom_segment()}} \tab \code{\link[=geom_textsegment]{geom_textsegment()}} \tab \code{\link[=geom_labelsegment]{geom_labelsegment()}} \cr
   \code{\link[ggplot2:geom_segment]{geom_curve()}} \tab \code{\link{geom_textcurve}} \tab \code{\link[=geom_labelcurve]{geom_labelcurve()}} \cr
   \code{\link[ggplot2:geom_abline]{geom_abline()}} \tab \code{\link[=geom_textabline]{geom_textabline()}} \tab \code{\link[=geom_labelabline]{geom_labelabline()}} \cr
   \code{\link[ggplot2:geom_abline]{geom_hline()}} \tab \code{\link[=geom_texthline]{geom_texthline()}} \tab \code{\link[=geom_labelhline]{geom_labelhline()}} \cr
   \code{\link[ggplot2:geom_abline]{geom_vline()}} \tab \code{\link[=geom_textvline]{geom_textvline()}} \tab \code{\link[=geom_labelvline]{geom_labelvline()}} \cr
   \code{\link[ggplot2:geom_density]{geom_density()}} \tab \code{\link[=geom_textdensity]{geom_textdensity()}} \tab \code{\link[=geom_labeldensity]{geom_labeldensity()}} \cr
   \code{\link[ggplot2:geom_smooth]{geom_smooth()}} \tab \code{\link[=geom_textsmooth]{geom_textsmooth()}} \tab \code{\link[=geom_labelsmooth]{geom_labelsmooth()}} \cr
   \code{\link[ggplot2:geom_contour]{geom_contour()}} \tab \code{\link[=geom_textcontour]{geom_textcontour()}} \tab \code{\link[=geom_labelcontour]{geom_labelcontour()}} \cr
   \code{\link[ggplot2:geom_density_2d]{geom_density2d()}} \tab \code{\link[=geom_textdensity2d]{geom_textdensity2d()}} \tab \code{\link[=geom_labeldensity2d]{geom_labeldensity2d()}} \cr
   \code{\link[ggplot2:ggsf]{geom_sf()}} \tab \code{\link[=geom_textsf]{geom_textsf()}} \tab \code{\link[=geom_labelsf]{geom_labelsf()}} \cr
}
}
