% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{waterways}
\alias{waterways}
\title{A simple features data frame of three Scottish waterways}
\format{
A data frame with 4 rows and 3 variables
\describe{
  \item{name}{name of geographic object}
  \item{type}{type of geographic object}
  \item{geometry}{sfc objects}
  }
}
\usage{
waterways
}
\description{
An `sf` object showing the River Clyde, the River Forth, and the canal
joining the two.
}
\details{
Contains Ordnance Survey data © Crown copyright and database right 2010-19
}
\keyword{datasets}
