% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomtextpath-package.R
\docType{package}
\name{geomtextpath-package}
\alias{geomtextpath}
\alias{geomtextpath-package}
\title{geomtextpath: Curved Text in 'ggplot2'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A 'ggplot2' extension that allows text to follow curved paths. Curved text makes it easier to directly label paths or neatly annotate in polar co-ordinates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://allancameron.github.io/geomtextpath/}
  \item Report bugs at \url{https://github.com/AllanCameron/geomtextpath/issues}
}

}
\author{
\strong{Maintainer}: Allan Cameron \email{Allan.Cameron@nhs.scot}

Authors:
\itemize{
  \item Teun van den Brand \email{tahvdbrand@gmail.com} (\href{https://orcid.org/0000-0002-9335-7468}{ORCID})
}

}
\keyword{internal}
