% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tetxpath.R
\docType{data}
\name{GeomTextPath}
\alias{GeomTextPath}
\title{The Geom object for a textpath}
\description{
This is the \code{ggproto} class that creates the textpath layer. It is not
intended to be used directly by the end user.
}
\keyword{datasets}
