% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_textpath.R, R/geom_labelpath.R,
%   R/geom_textsegment.R, R/geom_textabline.R, R/geom_textcontour.R,
%   R/geom_textcurve.R, R/geom_textdensity.R, R/geom_textdensity2d.R,
%   R/geom_texthline.R, R/geom_textsf.R
\docType{data}
\name{GeomTextpath}
\alias{GeomTextpath}
\alias{GeomTextline}
\alias{GeomLabelpath}
\alias{GeomLabelline}
\alias{GeomTextsegment}
\alias{GeomLabelsegment}
\alias{GeomTextabline}
\alias{GeomLabelabline}
\alias{GeomTextcontour}
\alias{GeomLabelcontour}
\alias{StatTextcontour}
\alias{GeomTextcurve}
\alias{GeomLabelcurve}
\alias{GeomTextdensity}
\alias{GeomLabeldensity}
\alias{GeomTextdensity2d}
\alias{GeomLabeldensity2d}
\alias{GeomTexthline}
\alias{GeomLabelhline}
\alias{GeomTextsf}
\alias{GeomLabelsf}
\title{The Geom object for a textpath}
\description{
This is the \code{ggproto} class that creates the textpath layer. It is not
intended to be used directly by the end user.
}
\keyword{datasets}
