% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_textpath.R
\name{textpathGrob}
\alias{textpathGrob}
\title{Draw text on a path.}
\usage{
textpathGrob(
  label,
  x = 0.5,
  y = 0.5,
  id = 1L,
  just = "centre",
  hjust = NULL,
  vjust = NULL,
  gp_text = gpar(),
  gp_path = gpar(),
  cut_path = NA,
  flip_inverted = FALSE,
  default.units = "npc",
  name = NULL,
  vp = NULL
)
}
\arguments{
\item{label}{A \code{character} vector.}

\item{x}{A \code{numeric} vector.}

\item{y}{A \code{numeric} vector.}

\item{id}{A \code{numeric} vector used to separate locations in \code{x} and \code{y} into
multiple lines. All locations with the same \code{id} belong to the same line.}

\item{just}{The justification of the text
    relative to its (x, y) location.  If there are two values, the first
    value specifies horizontal justification and the second value specifies
    vertical justification.  Possible string values are: \code{"left"},
    \code{"right"}, \code{"centre"}, \code{"center"}, \code{"bottom"},
    and \code{"top"}.  For numeric values, 0 means left (bottom) alignment
    and 1 means right (top) alignment.
  }

\item{hjust}{A numeric vector specifying horizontal justification.
    If specified, overrides the \code{just} setting.}

\item{vjust}{A numeric vector specifying vertical justification.
    If specified, overrides the \code{just} setting.}

\item{gp_text, gp_path}{An object of class \code{"gpar"}, typically the output from
a call from the \code{\link[grid:gpar]{gpar()}} function. These are basically lists
of graphical parameters for the text and path respectively.}

\item{cut_path}{A single logical TRUE or FALSE which if TRUE breaks the path
into two sections, one on either side of the string and if FALSE leaves the
path unbroken. The default value is NA, which will break the line if the
string has a vjust of between 0 and 1}

\item{flip_inverted}{If TRUE, any string where the majority of letters would
be upside down along the path are inverted to improve legibility. The
default is FALSE.}

\item{default.units}{A string indicating the default units to use
    if \code{x} or \code{y}
    are only given as numeric vectors.}

\item{name}{ A character identifier. }

\item{vp}{A Grid viewport object (or NULL).}
}
\value{
An object of class \code{gTree}, containing grobs.
}
\description{
This function creates (curved) text on a path.
}
\examples{
require(grid)

t <- seq(0, 2 * pi, length.out = 100)
grob <- textpathGrob(
  label = c(
    "Why I am making trigonometry jokes? Cos I can!",
    "I was never any good at sine language."
  ),
  x = c(t, t) / (2 * pi),
  y = c(cos(t), sin(t)) * 0.25 + 0.5,
  id = rep(1:2, each = length(t)),
  vjust = rep(0.5, 2 * length(t)),
  gp_text = gpar(lineheight = c(1.2, 1.2), fontsize = c(10, 10)),
  gp_path = gpar(lty = c(1, 2))
)

grid.newpage(); grid.draw(grob)
}
